Fix compilation on 64 bit systems

+++ equirect2cubemap.cpp
@@ -12,7 +12,7 @@ double fetch(T const* data, const ssize_t width, const ssize_t height,
              const size_t subpixelIndex, const ssize_t requestedX, const ssize_t requestedY)
 {
     const auto x = (requestedX+width) % width;
-    const auto y = std::clamp(requestedY, 0, height-1);
+    const auto y = std::clamp(requestedY, ssize_t(0), height-1);
 
     const auto rawValue = data[x*channelsPerPixel + subpixelIndex + y*rowStride];
